Created /api/player route which takes a name parameter. It then returns the players information, and their 10 most recent games.
